Init initializing and architecture
