+++ package.json
@@ -35,11 +35,16 @@
     "enzyme": "^2.9.1",
     "jest": "^20.0.4",
     "nodemon": "^1.11.0",
-    "react-test-renderer": "^15.6.1"
+    "react-test-renderer": "^15.6.1",
+    "webpack-merge": "^4.1.0"
   },
   "dependencies": {
+    "css-loader": "^0.28.4",
     "react": "^15.6.1",
     "react-dom": "^15.6.1",
-    "styled-components": "^2.1.2"
+    "sass-loader": "^6.0.6",
+    "style-loader": "^0.18.2",
+    "styled-components": "^2.1.2",
+    "webpack": "^3.5.2"
   }
 }
