+++ stories/index.js
@@ -1,6 +1,6 @@
 import { configure } from '@storybook/react';
 
-const req = require.context('../components', true, /\.story\.js$/)
+const req = require.context('../src/components', true, /\.story\.js$/)
 
 function loadStories() {
   req.keys().forEach((filename) => req(filename))
